s/float/double/g

+++ Canvas.h
@@ -4,10 +4,10 @@
 class Canvas
 {
 public:
-	virtual void setColour(float r, float g, float b) = 0;
-	virtual void moveTo(float x, float y) = 0;
-	virtual void lineTo(float x, float y) = 0;
-	virtual void circle(float x, float y, float r) = 0;
+	virtual void setColour(double r, double g, double b) = 0;
+	virtual void moveTo(double x, double y) = 0;
+	virtual void lineTo(double x, double y) = 0;
+	virtual void circle(double x, double y, double r) = 0;
 	virtual void stroke() = 0;
 	virtual void fill() = 0;
 };
